customer-service/requirements.txt
Flask==2.3.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
PyJWT==2.8.0
python-json-logger==2.0.7
prometheus-flask-exporter==0.23.0
python-logstash==0.4.8
pika==1.3.2
requests==2.31.0
Werkzeug==2.3.0